+++ CHANGES.md
@@ -19,6 +19,7 @@
 ## Minor code helpers
 
 * [SSHD-1040](https://issues.apache.org/jira/browse/SSHD-1040) Make server key available after KEX completed.
+* [SSHD-1030](https://issues.apache.org/jira/browse/SSHD-1030) Added a NoneFileSystemFactory implementation
 
 ## Behavioral changes and enhancements